requirements.txt
biopython==1.83
certifi==2024.2.2
charset-normalizer==3.3.2
gunc==1.0.6
idna==3.6
numpy==1.26.4
packaging==23.2
pandas==2.2.0
plotly==5.18.0
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
scipy==1.12.0
six==1.16.0
tenacity==8.2.3
tzdata==2023.4
urllib3==2.2.0
